VehicleTracker/requirements.txt
absl-py==0.9.0
astor==0.8.1
cycler==0.10.0
gast==0.4.0
google-pasta==0.2.0
grpcio==1.31.0
h5py==2.10.0
importlib-metadata==1.7.0
imutils==0.4.6
joblib==0.16.0
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
Markdown==3.2.2
matplotlib==3.1.0
mock==4.0.2
numpy==1.16.4
opencv-python==3.4.2.17
Pillow==6.0.0
protobuf==3.12.4
pyparsing==2.4.7
python-dateutil==2.8.1
PyYAML==5.3.1
scikit-learn==0.20.1
scipy==1.3.0
six==1.15.0
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
threadpoolctl==2.1.0
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0